requirements.txt
discord.py==1.7.3
SQLAlchemy==1.4.39
